initial run loop
