--- tools/nymvisor/src/upgrades/http_upstream.rs
@@ -1,2 +1,0 @@
-// Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
-// SPDX-License-Identifier: Apache-2.0
